# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SikOgami API",
    "description": "SikOgami \u2014 Fold. Breathe. Level Up. 30-level zen origami with Gemini Vision scan verification. Developer resources for SikOgami.\n\nVersioning: this API uses URL path versioning (/api/v1/...). The legacy unversioned /api/* paths remain as aliases. Breaking changes get a new path version (/api/v2/...); deprecations are announced in /CHANGELOG.md and signalled with Sunset and Deprecation response headers at least 90 days before removal.\n\nRate limits: 120 requests per 60 seconds per client. Responses include RateLimit-Policy, RateLimit, X-RateLimit-Limit, X-RateLimit-Remaining and X-RateLimit-Reset headers; 429 responses include Retry-After.\n\nErrors: every 4xx/5xx response returns the typed Error schema (ok:false, code, message, hint, status).",
    "version": "2.1.0",
    "contact": {
      "name": "SikOgami",
      "url": "https://sikogami.vercel.app",
      "email": "hello@sikogami.vercel.app"
    }
  },
  "servers": [
    {
      "url": "https://sikogami.vercel.app",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/v1/verify": {
      "post": {
        "summary": "SikOgami verify origami photo",
        "operationId": "verifyOrigami",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "image",
                  "levelId"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "description": "data:image/jpeg;base64,..."
                  },
                  "levelId": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 30
                  },
                  "levelTitle": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Always 200 with pass/score (mock, gemini or fallback mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Verify an origami photo via Gemini Vision. Send base64 data URL and levelId 1-30. The prompt is zen-generous (threshold 75). Without GEMINI_API_KEY the endpoint runs in mock mode (score 88+). Always 200 with pass/score/feedback/mode/expected.",
        "tags": [
          "verify"
        ]
      }
    },
    "/api/v1/progress": {
      "get": {
        "summary": "Get SikOgami progress",
        "operationId": "getProgress",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completed level ids for the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          },
          "400": {
            "description": "email required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Fetch completed level ids for a user by email query. Returns progress int[] 1-30, empty if user not found. Supports markdown content negotiation via Accept: text/markdown.",
        "tags": [
          "progress"
        ]
      },
      "post": {
        "summary": "Sync SikOgami progress",
        "operationId": "syncProgress",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "levelId": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 30
                  },
                  "progress": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 30
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated progress after sync",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing email, invalid levelId, or missing payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Sync level progress. Send {email, levelId} for single level or {email, progress:[1,2]} for bulk replace (DELETE+INSERT). Validates levelId 1-30. Returns updated progress array.",
        "tags": [
          "progress"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Typed error returned for every 4xx and 5xx response",
        "required": [
          "ok",
          "error",
          "code",
          "message",
          "status"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "description": "Human-readable error summary"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code",
            "example": "MISSING_FIELDS"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message"
          },
          "hint": {
            "type": "string",
            "description": "Resolution hint for the caller"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 400
          },
          "docs": {
            "type": "string",
            "format": "uri",
            "description": "Link to API documentation"
          }
        }
      },
...
      "ProgressResponse": {
        "type": "object",
        "required": [
          "ok",
          "progress"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "fallback": {
            "type": "boolean"
          },
          "progress": {
            "type": "array",
            "description": "Completed level ids (1-30)",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 30
            }
          }
        }
      },
      "VerifyResponse": {
        "type": "object",
        "required": [
          "pass",
          "score"
        ],
        "properties": {
          "pass": {
            "type": "boolean"
          },
          "score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 99
          },
          "feedback": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "gemini",
              "mock",
              "fallback",
              "error-fallback"
            ]
          },
          "expected": {
            "type": "string",
            "description": "Description of the expected origami model"
          }
        }
      },
...
    },
...
  },
...
}